Refactored Pam #156
