+++ client/src/components/pam/ExpansionItem.vue
