--- client/src/components/pam/Panel3.vue
@@ -1,15 +1,0 @@
-<template>
-  <q-list>
-    <!----- Quick ----->
-    <!-- <q-item
-      clickable
-      v-ripple
-      @click="onClickAll()"
-      active-class="bg-teal-1 text-grey-8"
-    >
-      <q-item-section>
-        <q-item-label lines="1">All</q-item-label>
-      </q-item-section>
-    </q-item> -->
-  </q-list>
-</template>
