+++ client/src/pages/Project.vue
@@ -248,16 +248,16 @@ import { date } from 'quasar'
 import { gql } from '@apollo/client'
 import { call } from 'vuex-pathify'
 
-import CitationBuilder from '@/components/shared/CitationBuilder'
-import AddFunding from '@/components/shared/modals/AddFunding'
-import AddLinks from '@/components/shared/modals/AddLinks'
-import FileManager from '@/components/fileManager/FileManager'
 import TextArea from '@/components/project/TextArea'
 import Header from '@/components/project/Header'
 import Links from '@/components/project/Links'
 import Funding from '@/components/project/Funding'
 import Collaborators from '@/components/project/Collaborators'
-import CollaboratorsModal from '@/components/shared/modals/Collaborators'
+import CollaboratorsModal from '@/components/project/modals/Collaborators'
+import AddFunding from '@/components/project/modals/AddFunding'
+import AddLinks from '@/components/project/modals/AddLinks'
+import FileManager from '@/components/fileManager/FileManager'
+import CitationBuilder from '@/components/shared/CitationBuilder'
 
 const defaullDescription = 'View Description'
